7.23.2
v7.22 => v7.23
- updated to NPOI-2.7.4;
- _GetImages() upgraded;

+++ Npoi/Excel.Table.column.cs
@@ -549,7 +549,7 @@ namespace Cliver
 
                 public int GetWidth()
                 {
-                    return Table._.GetColumnWidth(X - 1);
+                    return (int)Table._.GetColumnWidth(X - 1);
                 }
 
                 public void SetWidth(int width)
